feat(ui): add new location form component
- Introduce a new LocationForm component for handling location input and editing
- Add support for location submission, validation, and error handling
- Implement coordinate selection via map interaction and search
- Add loading states and form validation
- Include success/error feedback and navigation handling

refactor: update location handling and related components

- Replace old location form implementation with new LocationForm component
- Improve location page routing and navigation
- Add support for location editing and updates
- Enhance error handling and form validation
- Restructure location-related code and state management

feat(api): add location update endpoint

- Add new PUT endpoint for updating existing locations
- Implement slug-based updates with validation
- Add conflict checking for duplicate location names
- Support authenticated updates via CSRF protection

style: clean up old form implementation

- Remove obsolete code from dashboard/add.vue
- Replace with LocationForm component
- Improve code organization and readability

build: update dependencies and imports

- Add new imports and dependency management
- Update schema references and type definitions
- Ensure proper TypeScript support across components
