+++ lib/db/queries/location.ts
@@ -75,3 +75,11 @@ export async function insertLocation(insertable: InsertLocation, slug: string, u
   }).returning();
   return created;
 }
+
+export async function updateLocationBySlug(updates: InsertLocation, slug: string, userId: number) {
+  const [updated] = await db.update(location).set(updates).where(and(
+    eq(location.slug, slug),
+    eq(location.userId, userId),
+  )).returning();
+  return updated;
+}
