+++ pages/dashboard/location/[slug].vue
@@ -1,8 +1,14 @@
 <script lang="ts" setup>
 const locationStore = useLocationsStore();
+const route = useRoute();
 const { currentLocation: location, currentLocationStatus: status, currentLocationError: error } = storeToRefs(locationStore);
-onMounted(async () => {
+onMounted(() => {
   locationStore.refreshCurrentLocation();
+});
+onBeforeRouteUpdate((to) => {
+  if (to.name === "dashboard-location-slug") {
+    locationStore.refreshCurrentLocation();
+  }
 });
 </script>
 
@@ -11,7 +17,12 @@ onMounted(async () => {
     <div v-if="status === 'pending'">
       <div class="loading loading-spinner loading-xl" />
     </div>
-    <div v-if="location && status !== 'pending'">
+    <div v-if="error && status !== 'pending'" class="alert alert-error w-64">
+      <h2 class="text-lg">
+        {{ error.statusMessage }}
+      </h2>
+    </div>
+    <div v-if="route.name === 'dashboard-location-slug' && location && status !== 'pending'">
       <h2 class="text-xl">
         {{ location?.name }}
       </h2>
@@ -28,10 +39,10 @@ onMounted(async () => {
         </button>
       </div>
     </div>
-    <div v-if="error && status !== 'pending'" class="alert alert-error w-64">
-      <h2 class="text-lg">
-        {{ error.statusMessage }}
-      </h2>
+    <div
+      v-if="route.name !== 'dashboard-location-slug'"
+    >
+      <NuxtPage />
     </div>
   </div>
 </template>
