+++ server/api/locations/[slug].put.ts
@@ -1,0 +1,24 @@
+import { findLocationByName, updateLocationBySlug } from "~/lib/db/queries/location";
+import { InsertLocation } from "~/lib/db/schema";
+import defineAuthenticatedEventHandler from "~/utils/define-authenticated-event-handler";
+import SendZodError from "~/utils/send-zod-error";
+
+export default defineAuthenticatedEventHandler(async (event) => {
+  const slug = getRouterParam(event, "slug") as string;
+  const result = await readValidatedBody(event, InsertLocation.safeParse);
+
+  if (!result.success) {
+    return SendZodError(event, result.error);
+  }
+
+  const existingLocation = await findLocationByName(result.data, event.context.user.id);
+
+  if (existingLocation && existingLocation.slug !== slug) {
+    return sendError(event, createError({
+      statusCode: 409,
+      statusMessage: "A location with that name already exists. Please choose a different name.",
+    }));
+  }
+
+  return updateLocationBySlug(result.data, slug, event.context.user.id);
+});
